btrfs: scrub: requires target block group to be read-only if it's RAID56
Currently if a block group can not be marked read-only, we still
continue the scrub.

This is fine for all other mirrors based profiles, but can be a source
of problems for RAID56 P/Q scrubbing.

The scrub repair for data stripes would write back the repaired data
directly to the target disk, bypassing all the RAID56 RMW.
(This is the same for both the old and new scrub code).

But if RMW happened during that repair, we can have race between the RMW
and scrub data repair.

Thus here we take an preventive method to require the RAID56 block groups to be
marked read-only for scrub.

Signed-off-by: Qu Wenruo <wqu@suse.com>