btrfs: only run extent map shrinker inside cleaner kthread
There are still reports about extent map shrinker is causing high CPU
usage.

It turns out that the call backs can be very frequent, even inside
kswapd (and we can have multiple kswapd if the system have several
nodes).

For the only other fs implementing the reclaim callbacks, XFS does it
very differently by ensure the following conditions:

- Make sure there is only one reclaim work queued

- Add a delay before queuing the reclaim workload
  The default delay is 18s (60% of xfs_syncd_centisecs)

In btrfs, there is already a context which is very similar to the XFS
condition: cleaner kthread.

There is only one cleaner kthread for the fs, and it's waken periodically
(the same time interval as commit interval, which is 30s by default).

So it's much better to run the extent map shrinker inside cleaner
kthread.

And make the free_cached_objects() callback to only record the latest
number to free.

Link: https://lore.kernel.org/linux-btrfs/3df4acd616a07ef4d2dc6bad668701504b412ffc.camel@intelfx.name/
Link: https://lore.kernel.org/linux-btrfs/c30fd6b3-ca7a-4759-8a53-d42878bf84f7@gmail.com/
Reported-by: Ivan Shapovalov <intelfx@intelfx.name>
Reported-by: Jannik Glückert <jannik.glueckert@gmail.com>
Signed-off-by: Qu Wenruo <wqu@suse.com>
---
Reason for RFC:

I do not have an environment which can trigger the reclaim that
frequently, thus more tests would be appreciated.