btrfs: avoid unnecessary read for sectorsize < PAGE_SIZE case
[BUG]
The test case generic/573 always fails on btrfs if the sectorsize is
4K and page size is 64K:

    --- tests/generic/563.out	2024-04-25 18:13:45.178550333 +0930
    +++ /home/adam/xfstests-dev/results//generic/563.out.bad	2024-09-30 08:09:59.760877258 +0930
    @@ -3,7 +3,8 @@
     read is in range
     write is in range
     write -> read/write
    -read is in range
    +read has value of 8421376
    +read is NOT in range -33792 .. 33792
     write is in range

[CAUSE]
The read value is around 8M, which means btrfs has read the original 8M
file.

During buffered write, if btrfs detects a write doesn't cover a full
page, it will read out the full page.

However that is the legacy left from the old sectorsize == PAGE_SIZE
requirement, with the new sectorsize < PAGE_SIZE cases we can have a
case that the write range covers a full sector, but not a full page.

Then btrfs will trigger the read on the full page, resulting reading the
full file even we're re-writing the full page without any unaligned
write.

[FIX]
Fix prepare_uptodate_page() by passing the whole write range, and checks
if the write range is not aligned, and read the page only really needed.

This fix also slightly enhance the readability by moving the range
clamping into prepare_uptodate_page(), so we do not need to treat the
leading and tailing part differently.

Signed-off-by: Qu Wenruo <wqu@suse.com>