btrfs: return error if the primary super block write to a new device failed
When adding a new device to btrfs, btrfs adds the device and relying on
the regular transaction commit to put a new super block to that device.

However during that transaction commit, if the primary super block write
back failed for that device, the fs will not be able to be mounted
unless degraded.

This is because the primary super block failed to be written to that new
device, thus btrfs is unable to locate that new device thus it will be
treated as missing.

This means for the initial transaction commit on that new device, it is
more critical than other devices.

Treat primary super block write back to a new device as a critical error,
so that if that error happened, we will abort the transaction.

Although transaction abort is annoying, it keeps the fs to be mountable.

Signed-off-by: Qu Wenruo <wqu@suse.com>